pyproject.toml
[project]
name = "documenters-cle-langchain"
version = "0.1.0"
description = "LangChain scaffold for Google Drive document routing and extraction"
readme = "README.md"
authors = [
    { name = "David Eads", email = "davideads@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.12.5",
    "google-auth>=2.0",
    "google-api-python-client>=2.0",
    "python-dotenv>=1.0",
]

[project.scripts]
documenters-cle-langchain = "documenters_cle_langchain.cli:main"

[build-system]
requires = ["uv_build>=0.9.11,<0.10.0"]
build-backend = "uv_build"
